SIG: Move ingress to internal
