+++ go/sig/BUILD.bazel
@@ -15,9 +15,9 @@ go_library(
         "//go/lib/serrors:go_default_library",
         "//go/sig/config:go_default_library",
         "//go/sig/egress:go_default_library",
-        "//go/sig/ingress:go_default_library",
         "//go/sig/internal/base:go_default_library",
         "//go/sig/internal/disp:go_default_library",
+        "//go/sig/internal/ingress:go_default_library",
         "//go/sig/internal/metrics:go_default_library",
         "//go/sig/internal/sigconfig:go_default_library",
         "//go/sig/internal/xnet:go_default_library",
