+++ go/sig/internal/ingress/dispatcher.go
