+++ go/sig/internal/ingress/worker.go
